dashboard/requirements.txt
streamlit
streamlit-autorefresh
psycopg2-binary
pandas
neo4j
